notifications/notification_api_flask/requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
Flask==2.3.3
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-Mobility==1.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.9.0
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.29.0
oauthlib==3.2.2
pika==1.3.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
python-logstash==0.4.8
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sentry-sdk==1.40.5
six==1.16.0
SQLAlchemy==2.0.27
urllib3==2.2.1
Werkzeug==2.3.8
zope.event==5.0
zope.interface==6.1
marshmallow-dataclass==8.7.1